git-svn-id: http://kos.informatik.uni-osnabrueck.de/software/meshing@2389 bea721bf-f9f1-0310-a2d5-e0bfdc8dd81e

+++ src/viewer/EventHandler.cpp
@@ -122,7 +122,7 @@ void EventHandler::transform_from_file(){
 	//Load matrix
 	ifstream in(filename.toStdString().c_str());
 
-	double* matrix_data = new double[16];
+	float* matrix_data = new float[16];
 
 	for(int i = 0; i < 16; i++){
 		if(!in.good()){
